feat: add theme switch container
