+++ src/layouts/Header/ThemeSwitcher/ThemeSwitcher.jsx
@@ -1,20 +1,11 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
 
 import styles from './ThemeSwitcher.module.scss';
 import { THEMES } from '../../../constants/themes';
 import { darkIcon } from '../../../assets/svg/dark-theme-icon.js';
 import { lightIcon } from '../../../assets/svg/light-theme-icon.js';
-import { setCurrentTheme } from '../../../redux/actions/themeActions';
 
-function ThemeSwitcher({ theme, switchTheme }) {
-  function onSwitch() {
-    const currentTheme = theme === THEMES.light ? THEMES.dark : THEMES.light;
-
-    switchTheme(currentTheme);
-  }
-
+function ThemeSwitcher({ theme, onSwitch }) {
   const lightThemeBtnClasses =
     theme === THEMES.light
       ? [styles.switchThemeBtn, styles.active, styles.lightBtn].join(' ')
@@ -33,14 +24,4 @@ function ThemeSwitcher({ theme, switchTheme }) {
   );
 }
 
-const mapStateToProps = state => ({
-  theme: state.theme.currentTheme
-});
-
-const mapDispatchToProps = dispatch => {
-  return {
-    switchTheme: theme => dispatch(setCurrentTheme(theme))
-  };
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(ThemeSwitcher);
+export default ThemeSwitcher;
